Remove deprecated config encoding
This will be a breaking change, since `configurableProvider` overrode `Configure` instead
of `CheckConfig`, JSON encoded strings will be present in old state.

Unless we are comfortable with the breaking (which would surprise me), we should upgrade
the version of pulumi-go-provider used but not remove `configurableProvider` until we are
comfortable making the breaking change.